scripts/requirements.txt
supabase
click
python-dotenv